Criando mapa com todas posicoes por id
