+++ src/pages/all-equipaments-map/Map.css
