+++ src/pages/equipament-position-map/Map.css
@@ -1,0 +1,4 @@
+.map-container  {
+    width: 1200px;
+    height: 1200px;
+}
